debugging update and add methods

+++ java-school-fullstack/src/main/java/mthree/com/fullstackschool/service/StudentServiceImpl.java
@@ -7,6 +7,7 @@ import mthree.com.fullstackschool.model.Course;
 import mthree.com.fullstackschool.model.Student;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
+import org.springframework.dao.DataAccessResourceFailureException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -17,10 +18,12 @@ public class StudentServiceImpl implements StudentServiceInterface {
     //YOUR CODE STARTS HERE
     private final JdbcTemplate jdbcTemplate;
     private final StudentDao studentDao;
+    private final CourseDao courseDao;
 
-    public StudentServiceImpl(JdbcTemplate jdbcTemplate, StudentDao studentDao) {
+    public StudentServiceImpl(JdbcTemplate jdbcTemplate, StudentDao studentDao, CourseDao courseDao) {
         this.jdbcTemplate = jdbcTemplate;
         this.studentDao = studentDao;
+        this.courseDao = courseDao;
     }
 
     //YOUR CODE ENDS HERE
@@ -35,15 +38,26 @@ public class StudentServiceImpl implements StudentServiceInterface {
 
     public Student getStudentById(int id) {
         //YOUR CODE STARTS HERE
-
-        return studentDao.findStudentById(id);
-
+        try {
+            return studentDao.findStudentById(id);
+        } catch (DataAccessException ex) {
+            Student studentNotFound = new Student();
+            studentNotFound.setStudentFirstName("Student NOT Found");
+            studentNotFound.setStudentLastName("Student NOT Found");
+            return studentNotFound;
+        }
         //YOUR CODE ENDS HERE
     }
 
     public Student addNewStudent(Student student) {
         //YOUR CODE STARTS HERE
 
+        if (student.getStudentFirstName() == null || student.getStudentFirstName().trim().isEmpty() ||
+            student.getStudentLastName() == null || student.getStudentLastName().trim().isEmpty()) {
+            student.setStudentFirstName("First Name blank, student NOT added");
+            student.setStudentLastName("Last Name blank, student NOT added");
+            return student;
+        }
         return studentDao.createNewStudent(student);
 
         //YOUR CODE ENDS HERE
@@ -55,6 +69,7 @@ public class StudentServiceImpl implements StudentServiceInterface {
         if ( id != student.getStudentId()) {
             student.setStudentFirstName("IDs do not match, student not updated");
             student.setStudentLastName("IDs do not match, student not updated");
+            return student;
         }
 
         studentDao.updateStudent(student);
@@ -74,7 +89,18 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public void deleteStudentFromCourse(int studentId, int courseId) {
         //YOUR CODE STARTS HERE
 
-        studentDao.deleteStudentFromCourse(studentId, courseId);
+        Student student = getStudentById(studentId);
+        Course course = courseDao.findCourseById(courseId);
+
+        if (student.getStudentFirstName().equals("Student Not Found")) {
+            System.out.println("Student Not Found");
+        } else if (course.getCourseName().equals("Course Not Found")) {
+            System.out.println("Course Not Found");
+        } else {
+            studentDao.deleteStudentFromCourse(studentId, courseId);
+            System.out.println("Student: " + studentId + " deleted from course: " + courseId);
+        }
+
 
         //YOUR CODE ENDS HERE
     }
@@ -82,7 +108,21 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public void addStudentToCourse(int studentId, int courseId) {
         //YOUR CODE STARTS HERE
 
-        studentDao.addStudentToCourse(studentId, courseId);
+        Student student = getStudentById(studentId);
+        Course course = courseDao.findCourseById(courseId);
+
+        try {
+            if (student.getStudentFirstName().equals("Student Not Found")) {
+                System.out.println("Student Not Found");
+            } else if (course.getCourseName().equals("Course Not Found")) {
+                System.out.println("Course Not Found");
+            } else {
+                studentDao.addStudentToCourse(studentId, courseId);
+                System.out.println("Student: " + studentId + " added from course: " + courseId);
+            }
+        } catch (DataAccessException ex) {
+            System.out.println("Student: " + studentId + " already enrolled in course: " + courseId);
+        }
 
         //YOUR CODE ENDS HERE
     }
